First Commit.
Here i only developed the bluetooth manager, and the first presentation page where all the BLE devices can be seen, however it has to perform better and also in combination with the other ContentViews.

Next Step is to connect it to the IMU and IADC. Please check the app in a device one more time before you change it. It is full of bugs and once you try it you'll know what to change.

-José Heinz